filter/fir.h: fix overlapping memcpy
Previously the setTaps method invoked UB by performing memcpy on
overlapping objects when _taps.size > oldTC.

This caused it to fail on hardened systems like Alpine, which use
e.g. http://git.2f30.org/fortify-headers/

Fixed by using std::copy, which is defined for when the
destination iterator is outside the range to be copied, even if
the ranges overlap.